+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -3623,6 +3623,16 @@ $ASYN_Completers = {
             break
         }
 
+        # Amazon.AppSync.GraphQLApiType
+        {
+            ($_ -eq "Get-ASYNGraphqlApiList/ApiType") -Or
+            ($_ -eq "New-ASYNGraphqlApi/ApiType")
+        }
+        {
+            $v = "GRAPHQL","MERGED"
+            break
+        }
+
         # Amazon.AppSync.GraphQLApiVisibility
         "New-ASYNGraphqlApi/Visibility"
         {
@@ -3630,10 +3640,28 @@ $ASYN_Completers = {
             break
         }
 
+        # Amazon.AppSync.MergeType
+        {
+            ($_ -eq "Start-ASYNMergedGraphqlApi/SourceApiAssociationConfig_MergeType") -Or
+            ($_ -eq "Start-ASYNSourceGraphqlApi/SourceApiAssociationConfig_MergeType") -Or
+            ($_ -eq "Update-ASYNSourceApiAssociation/SourceApiAssociationConfig_MergeType")
+        }
+        {
+            $v = "AUTO_MERGE","MANUAL_MERGE"
+            break
+        }
+
         # Amazon.AppSync.OutputType
         "Get-ASYNIntrospectionSchema/Format"
         {
             $v = "JSON","SDL"
+            break
+        }
+
+        # Amazon.AppSync.Ownership
+        "Get-ASYNGraphqlApiList/Owner"
+        {
+            $v = "CURRENT_ACCOUNT","OTHER_ACCOUNTS"
             break
         }
 
@@ -3674,6 +3702,7 @@ $ASYN_Completers = {
         {
             ($_ -eq "Get-ASYNType/Format") -Or
             ($_ -eq "Get-ASYNTypeList/Format") -Or
+            ($_ -eq "Get-ASYNTypesByAssociationList/Format") -Or
             ($_ -eq "New-ASYNType/Format") -Or
             ($_ -eq "Update-ASYNType/Format")
         }
@@ -3692,12 +3721,15 @@ $ASYN_Completers = {
 
 $ASYN_map = @{
     "ApiCachingBehavior"=@("New-ASYNApiCache","Update-ASYNApiCache")
+    "ApiType"=@("Get-ASYNGraphqlApiList","New-ASYNGraphqlApi")
     "AuthenticationType"=@("New-ASYNGraphqlApi","Update-ASYNGraphqlApi")
-    "Format"=@("Get-ASYNIntrospectionSchema","Get-ASYNType","Get-ASYNTypeList","New-ASYNType","Update-ASYNType")
+    "Format"=@("Get-ASYNIntrospectionSchema","Get-ASYNType","Get-ASYNTypeList","Get-ASYNTypesByAssociationList","New-ASYNType","Update-ASYNType")
     "Kind"=@("New-ASYNResolver","Update-ASYNResolver")
     "LogConfig_FieldLogLevel"=@("New-ASYNGraphqlApi","Update-ASYNGraphqlApi")
+    "Owner"=@("Get-ASYNGraphqlApiList")
     "RelationalDatabaseConfig_RelationalDatabaseSourceType"=@("New-ASYNDataSource","Update-ASYNDataSource")
     "Runtime_Name"=@("New-ASYNFunction","New-ASYNResolver","Test-ASYNCode","Update-ASYNFunction","Update-ASYNResolver")
+    "SourceApiAssociationConfig_MergeType"=@("Start-ASYNMergedGraphqlApi","Start-ASYNSourceGraphqlApi","Update-ASYNSourceApiAssociation")
     "SyncConfig_ConflictDetection"=@("New-ASYNFunction","New-ASYNResolver","Update-ASYNFunction","Update-ASYNResolver")
     "SyncConfig_ConflictHandler"=@("New-ASYNFunction","New-ASYNResolver","Update-ASYNFunction","Update-ASYNResolver")
     "Type"=@("New-ASYNApiCache","New-ASYNDataSource","Update-ASYNApiCache","Update-ASYNDataSource")
@@ -3755,6 +3787,8 @@ $ASYN_SelectCompleters = {
 
 $ASYN_SelectMap = @{
     "Select"=@("Start-ASYNApiAssociation",
+               "Start-ASYNMergedGraphqlApi",
+               "Start-ASYNSourceGraphqlApi",
                "New-ASYNApiCache",
                "New-ASYNApiKey",
                "New-ASYNDataSource",
@@ -3772,6 +3806,8 @@ $ASYN_SelectMap = @{
                "Remove-ASYNResolver",
                "Remove-ASYNType",
                "Stop-ASYNApiAssociation",
+               "Stop-ASYNMergedGraphqlApi",
+               "Stop-ASYNSourceGraphqlApi",
                "Test-ASYNCode",
                "Test-ASYNMappingTemplate",
                "Clear-ASYNApiCache",
@@ -3784,6 +3820,7 @@ $ASYN_SelectMap = @{
                "Get-ASYNIntrospectionSchema",
                "Get-ASYNResolver",
                "Get-ASYNSchemaCreationStatus",
+               "Get-ASYNSourceApiAssociation",
                "Get-ASYNType",
                "Get-ASYNApiKeyList",
                "Get-ASYNDataSourceList",
@@ -3792,9 +3829,12 @@ $ASYN_SelectMap = @{
                "Get-ASYNGraphqlApiList",
                "Get-ASYNResolverList",
                "Get-ASYNResolverListByFunction",
+               "Get-ASYNSourceApiAssociationList",
                "Get-ASYNResourceTag",
                "Get-ASYNTypeList",
+               "Get-ASYNTypesByAssociationList",
                "Start-ASYNSchemaCreation",
+               "Start-ASYNSchemaMerge",
                "Add-ASYNResourceTag",
                "Remove-ASYNResourceTag",
                "Update-ASYNApiCache",
@@ -3804,6 +3844,7 @@ $ASYN_SelectMap = @{
                "Update-ASYNFunction",
                "Update-ASYNGraphqlApi",
                "Update-ASYNResolver",
+               "Update-ASYNSourceApiAssociation",
                "Update-ASYNType")
 }
 
@@ -14446,7 +14487,7 @@ $CUR_Completers = {
             ($_ -eq "Write-CURReportDefinition/ReportDefinition_S3Region")
         }
         {
-            $v = "af-south-1","ap-east-1","ap-northeast-1","ap-northeast-2","ap-northeast-3","ap-south-1","ap-southeast-1","ap-southeast-2","ap-southeast-3","ca-central-1","cn-north-1","cn-northwest-1","eu-central-1","eu-north-1","eu-south-1","eu-south-2","eu-west-1","eu-west-2","eu-west-3","me-central-1","me-south-1","sa-east-1","us-east-1","us-east-2","us-west-1","us-west-2"
+            $v = "af-south-1","ap-east-1","ap-northeast-1","ap-northeast-2","ap-northeast-3","ap-south-1","ap-south-2","ap-southeast-1","ap-southeast-2","ap-southeast-3","ca-central-1","cn-north-1","cn-northwest-1","eu-central-1","eu-central-2","eu-north-1","eu-south-1","eu-south-2","eu-west-1","eu-west-2","eu-west-3","me-central-1","me-south-1","sa-east-1","us-east-1","us-east-2","us-west-1","us-west-2"
             break
         }