+++ modules/AWSPowerShell/Cmdlets/AppSync/Basic/Update-ASYNGraphqlApi-Cmdlet.cs
@@ -194,6 +194,19 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
         public System.String OpenIDConnectConfig_Issuer { get; set; }
         #endregion
         
+        #region Parameter MergedApiExecutionRoleArn
+        /// <summary>
+        /// <para>
+        /// <para>The Identity and Access Management service role ARN for a merged API. The AppSync
+        /// service assumes this role on behalf of the Merged API to validate access to source
+        /// APIs at runtime and to prompt the <code>AUTO_MERGE</code> to update the merged API
+        /// endpoint with the source API changes automatically.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String MergedApiExecutionRoleArn { get; set; }
+        #endregion
+        
         #region Parameter Name
         /// <summary>
         /// <para>
@@ -209,6 +222,16 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
         #endif
         [Amazon.PowerShell.Common.AWSRequiredParameter]
         public System.String Name { get; set; }
+        #endregion
+        
+        #region Parameter OwnerContact
+        /// <summary>
+        /// <para>
+        /// <para>The owner contact information for an API resource.</para><para>This field accepts any string input with a length of 0 - 256 characters.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String OwnerContact { get; set; }
         #endregion
         
         #region Parameter UserPoolConfig
@@ -311,6 +334,7 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
             context.LogConfig_CloudWatchLogsRoleArn = this.LogConfig_CloudWatchLogsRoleArn;
             context.LogConfig_ExcludeVerboseContent = this.LogConfig_ExcludeVerboseContent;
             context.LogConfig_FieldLogLevel = this.LogConfig_FieldLogLevel;
+            context.MergedApiExecutionRoleArn = this.MergedApiExecutionRoleArn;
             context.Name = this.Name;
             #if MODULAR
             if (this.Name == null && ParameterWasBound(nameof(this.Name)))
@@ -322,6 +346,7 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
             context.OpenIDConnectConfig_ClientId = this.OpenIDConnectConfig_ClientId;
             context.OpenIDConnectConfig_IatTTL = this.OpenIDConnectConfig_IatTTL;
             context.OpenIDConnectConfig_Issuer = this.OpenIDConnectConfig_Issuer;
+            context.OwnerContact = this.OwnerContact;
             context.UserPoolConfig = this.UserPoolConfig;
             context.XrayEnabled = this.XrayEnabled;
             
@@ -430,6 +455,10 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
             {
                 request.LogConfig = null;
             }
+            if (cmdletContext.MergedApiExecutionRoleArn != null)
+            {
+                request.MergedApiExecutionRoleArn = cmdletContext.MergedApiExecutionRoleArn;
+            }
             if (cmdletContext.Name != null)
             {
                 request.Name = cmdletContext.Name;
@@ -482,6 +511,10 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
             if (requestOpenIDConnectConfigIsNull)
             {
                 request.OpenIDConnectConfig = null;
+            }
+            if (cmdletContext.OwnerContact != null)
+            {
+                request.OwnerContact = cmdletContext.OwnerContact;
             }
             if (cmdletContext.UserPoolConfig != null)
             {
@@ -561,11 +594,13 @@ namespace Amazon.PowerShell.Cmdlets.ASYN
             public System.String LogConfig_CloudWatchLogsRoleArn { get; set; }
             public System.Boolean? LogConfig_ExcludeVerboseContent { get; set; }
             public Amazon.AppSync.FieldLogLevel LogConfig_FieldLogLevel { get; set; }
+            public System.String MergedApiExecutionRoleArn { get; set; }
             public System.String Name { get; set; }
             public System.Int64? OpenIDConnectConfig_AuthTTL { get; set; }
             public System.String OpenIDConnectConfig_ClientId { get; set; }
             public System.Int64? OpenIDConnectConfig_IatTTL { get; set; }
             public System.String OpenIDConnectConfig_Issuer { get; set; }
+            public System.String OwnerContact { get; set; }
             public Amazon.AppSync.Model.UserPoolConfig UserPoolConfig { get; set; }
             public System.Boolean? XrayEnabled { get; set; }
             public System.Func<Amazon.AppSync.Model.UpdateGraphqlApiResponse, UpdateASYNGraphqlApiCmdlet, object> Select { get; set; } =